+++ .github/ISSUE_TEMPLATE/showcase.md
@@ -1,0 +1,19 @@
+---
+name: Showcase
+about: Share something you built with Mesop
+title: ''
+labels: showcase
+assignees: ''
+---
+
+**Describe your showcase**
+A clear and concise description of what you built with Mesop.
+
+**GitHub Repo**
+The link to your GitHub repo or specific file.
+
+**Image**
+Upload an image or screenshot of your application:
+
+**Application URL** (if applicable)
+If your application is hosted online, provide the URL here:
