+++ docs/javascript/showcase.js
@@ -1,0 +1,4 @@
+const iframe = document.createElement('iframe');
+iframe.src = 'https://wwwillchen-mesop-showcase.hf.space/';
+iframe.className = 'full-screen-iframe';
+document.body.appendChild(iframe);
