+++ docs/showcase.md
@@ -1,0 +1,9 @@
+<!-- This file path is specifically chosen to match the GitHub pages path. -->
+---
+hide:
+  - navigation
+  - toc
+---
+<!-- This uses the production URLs to make the paths work locally and on GitHub -->
+<link rel="stylesheet" href="https://google.github.io/mesop/stylesheets/demo.css">
+<script src="https://google.github.io/mesop/javascript/showcase.js"></script>
