+++ showcase/README.md
@@ -1,0 +1,10 @@
+---
+title: Mesop Showcase
+emoji: 👓
+colorFrom: red
+colorTo: yellow
+sdk: docker
+pinned: false
+license: apache-2.0
+app_port: 8080
+---
