+++ showcase/requirements.txt
@@ -1,0 +1,2 @@
+mesop>=0.11.1
+gunicorn>=22.0.0
